requirements.txt
-e .
aws-cdk.aws-lambda
aws-cdk.aws-apigateway